+++ src/utils/fix-utils.js
@@ -1,0 +1,33 @@
+/**
+ * Copyright IBM Corp. 2022, 2022
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+export const fixUsingMap = (value, target, map) => {
+  let workingValue = value;
+  let match = target.exec(workingValue);
+
+  while (match) {
+    if (map[match[0]]) {
+      workingValue = workingValue.replace(match[0], map[match[0]]);
+    }
+
+    match = target.exec(workingValue);
+  }
+
+  return workingValue;
+};
+
+export const tryFix = ({ target, replacement, version }, value, ruleInfo) => {
+  if (version === undefined || ruleInfo.version.startsWith(version)) {
+    if (typeof replacement === "function") {
+      return replacement(value, target);
+    }
+
+    return value.replaceAll(target, replacement);
+  }
+
+  return value;
+};